Use the recommended plain setup.py installation for upstream
the setup.py is too opinionated with its builddir/install setup to fit
into our pattern here. Also, they want to erase allDict.cxx.pch there
during install and things like that so we cannot hook into the setup
without rewriting it completely which is not the point of "upstream".