Revert system_dirs.diff
this patch from upstream should only be applied on Linux builds.
However, probably because the tarball is generated on Linux, it's
applied in the source tarball already.

In any case, it is unclear whether this patch is needed for Linux in
conda-forge. See
https://github.com/conda-forge/cppyy-cling-feedstock/pull/40#issuecomment-975781900
and https://github.com/wlav/cppyy/issues/156 for what seems to be the
relevant discussion.

We unapply this patch here unconditionally since it is wrong on macOS
and it also complicates writing later patches here to simplify the logic
for conda-forge.